+++ packages/wiz-ui-next/src/components/base/menu/menu-item.stories.ts
@@ -122,3 +122,9 @@ FontSize.args = {
   label: "テスト",
   fontSize: "lg",
 };
+
+export const Selected = Template.bind({});
+Selected.args = {
+  label: "テスト",
+  selected: true,
+};