+++ packages/wiz-ui/src/components/base/inputs/search-input/search-input.vue
@@ -21,7 +21,7 @@
         "
         autocomplete="off"
       />
-      <WizISearch :class="searchInputIconStyle" />
+      <component :is="icon" :class="searchInputIconStyle" />
     </div>
     <WizPopup
       :isOpen="!disabled && openPopup"
@@ -149,7 +149,7 @@ import {
   WizPopupContainer,
   WizSearchPopup,
 } from "@/components";
-import { WizICheck, WizIChevronRight } from "@/components/icons";
+import { WizICheck, WizIChevronRight, TIcon } from "@/components/icons";
 
 import { SearchInputOption } from "./types";
 
@@ -199,6 +199,11 @@ const props = defineProps({
     type: Boolean,
     required: false,
     default: false,
+  },
+  icon: {
+    type: Object as PropType<TIcon>,
+    required: false,
+    default: WizISearch,
   },
 });